+++ packages/atproto/lib/src/entities/repo_op.freezed.dart
@@ -20,8 +20,14 @@ RepoOp _$RepoOpFromJson(Map<String, dynamic> json) {
 
 /// @nodoc
 mixin _$RepoOp {
+  @Uint8ListConverter()
+  Uint8List? get cid => throw _privateConstructorUsedError;
   RepoAction get action => throw _privateConstructorUsedError;
+  @Deprecated('Use uri instead. Will be removed in v0.6.0')
   String get path => throw _privateConstructorUsedError;
+  @AtUriConverter()
+  AtUri get uri => throw _privateConstructorUsedError;
+  Map<String, dynamic>? get record => throw _privateConstructorUsedError;
 
   Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
   @JsonKey(ignore: true)
@@ -33,7 +39,12 @@ abstract class $RepoOpCopyWith<$Res> {
   factory $RepoOpCopyWith(RepoOp value, $Res Function(RepoOp) then) =
       _$RepoOpCopyWithImpl<$Res, RepoOp>;
   @useResult
-  $Res call({RepoAction action, String path});
+  $Res call(
+      {@Uint8ListConverter() Uint8List? cid,
+      RepoAction action,
+      @Deprecated('Use uri instead. Will be removed in v0.6.0') String path,
+      @AtUriConverter() AtUri uri,
+      Map<String, dynamic>? record});
 }
 
 /// @nodoc
@@ -49,10 +60,17 @@ class _$RepoOpCopyWithImpl<$Res, $Val extends RepoOp>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? cid = freezed,
     Object? action = null,
     Object? path = null,
+    Object? uri = null,
+    Object? record = freezed,
   }) {
     return _then(_value.copyWith(
+      cid: freezed == cid
+          ? _value.cid
+          : cid // ignore: cast_nullable_to_non_nullable
+              as Uint8List?,
       action: null == action
           ? _value.action
           : action // ignore: cast_nullable_to_non_nullable
@@ -61,6 +79,14 @@ class _$RepoOpCopyWithImpl<$Res, $Val extends RepoOp>
           ? _value.path
           : path // ignore: cast_nullable_to_non_nullable
               as String,
+      uri: null == uri
+          ? _value.uri
+          : uri // ignore: cast_nullable_to_non_nullable
+              as AtUri,
+      record: freezed == record
+          ? _value.record
+          : record // ignore: cast_nullable_to_non_nullable
+              as Map<String, dynamic>?,
     ) as $Val);
   }
 }
@@ -71,7 +97,12 @@ abstract class _$$_RepoOpCopyWith<$Res> implements $RepoOpCopyWith<$Res> {
       __$$_RepoOpCopyWithImpl<$Res>;
   @override
   @useResult
-  $Res call({RepoAction action, String path});
+  $Res call(
+      {@Uint8ListConverter() Uint8List? cid,
+      RepoAction action,
+      @Deprecated('Use uri instead. Will be removed in v0.6.0') String path,
+      @AtUriConverter() AtUri uri,
+      Map<String, dynamic>? record});
 }
 
 /// @nodoc
@@ -84,10 +115,17 @@ class __$$_RepoOpCopyWithImpl<$Res>
   @pragma('vm:prefer-inline')
   @override
   $Res call({
+    Object? cid = freezed,
     Object? action = null,
     Object? path = null,
+    Object? uri = null,
+    Object? record = freezed,
   }) {
     return _then(_$_RepoOp(
+      cid: freezed == cid
+          ? _value.cid
+          : cid // ignore: cast_nullable_to_non_nullable
+              as Uint8List?,
       action: null == action
           ? _value.action
           : action // ignore: cast_nullable_to_non_nullable
@@ -96,6 +134,14 @@ class __$$_RepoOpCopyWithImpl<$Res>
           ? _value.path
           : path // ignore: cast_nullable_to_non_nullable
               as String,
+      uri: null == uri
+          ? _value.uri
+          : uri // ignore: cast_nullable_to_non_nullable
+              as AtUri,
+      record: freezed == record
+          ? _value._record
+          : record // ignore: cast_nullable_to_non_nullable
+              as Map<String, dynamic>?,
     ));
   }
 }
@@ -103,19 +149,42 @@ class __$$_RepoOpCopyWithImpl<$Res>
 /// @nodoc
 @JsonSerializable()
 class _$_RepoOp implements _RepoOp {
-  const _$_RepoOp({required this.action, required this.path});
+  const _$_RepoOp(
+      {@Uint8ListConverter() this.cid,
+      required this.action,
+      @Deprecated('Use uri instead. Will be removed in v0.6.0')
+      required this.path,
+      @AtUriConverter() required this.uri,
+      final Map<String, dynamic>? record})
+      : _record = record;
 
   factory _$_RepoOp.fromJson(Map<String, dynamic> json) =>
       _$$_RepoOpFromJson(json);
 
   @override
+  @Uint8ListConverter()
+  final Uint8List? cid;
+  @override
   final RepoAction action;
   @override
+  @Deprecated('Use uri instead. Will be removed in v0.6.0')
   final String path;
+  @override
+  @AtUriConverter()
+  final AtUri uri;
+  final Map<String, dynamic>? _record;
+  @override
+  Map<String, dynamic>? get record {
+    final value = _record;
+    if (value == null) return null;
+    if (_record is EqualUnmodifiableMapView) return _record;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableMapView(value);
+  }
 
   @override
   String toString() {
-    return 'RepoOp(action: $action, path: $path)';
+    return 'RepoOp(cid: $cid, action: $action, path: $path, uri: $uri, record: $record)';
   }
 
   @override
@@ -123,13 +192,22 @@ class _$_RepoOp implements _RepoOp {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
             other is _$_RepoOp &&
+            const DeepCollectionEquality().equals(other.cid, cid) &&
             (identical(other.action, action) || other.action == action) &&
-            (identical(other.path, path) || other.path == path));
+            (identical(other.path, path) || other.path == path) &&
+            (identical(other.uri, uri) || other.uri == uri) &&
+            const DeepCollectionEquality().equals(other._record, _record));
   }
 
   @JsonKey(ignore: true)
   @override
-  int get hashCode => Object.hash(runtimeType, action, path);
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(cid),
+      action,
+      path,
+      uri,
+      const DeepCollectionEquality().hash(_record));
 
   @JsonKey(ignore: true)
   @override
@@ -147,15 +225,28 @@ class _$_RepoOp implements _RepoOp {
 
 abstract class _RepoOp implements RepoOp {
   const factory _RepoOp(
-      {required final RepoAction action,
-      required final String path}) = _$_RepoOp;
+      {@Uint8ListConverter() final Uint8List? cid,
+      required final RepoAction action,
+      @Deprecated('Use uri instead. Will be removed in v0.6.0')
+      required final String path,
+      @AtUriConverter() required final AtUri uri,
+      final Map<String, dynamic>? record}) = _$_RepoOp;
 
   factory _RepoOp.fromJson(Map<String, dynamic> json) = _$_RepoOp.fromJson;
 
   @override
+  @Uint8ListConverter()
+  Uint8List? get cid;
+  @override
   RepoAction get action;
   @override
+  @Deprecated('Use uri instead. Will be removed in v0.6.0')
   String get path;
+  @override
+  @AtUriConverter()
+  AtUri get uri;
+  @override
+  Map<String, dynamic>? get record;
   @override
   @JsonKey(ignore: true)
   _$$_RepoOpCopyWith<_$_RepoOp> get copyWith =>