+++ packages/atproto_core/lib/src/client/auth_required_client.dart
@@ -92,6 +92,7 @@ class AuthRequiredClient implements Client {
     final String? service,
     final Map<String, dynamic>? parameters,
     final xrpc.To<T>? to,
+    final xrpc.JsonConverter? converter,
   }) =>
       throw UnimplementedError();
 }